Create Virus
Save it as “virus.bat” or anything with the name you want to. Now your first virus is ready to work. You can also convert it as .exe file. It copies itself to startup, hard disk, my documents and then it kills explorer.exe, chrome.exe and firefox.exe. It tries to delete mcafee antivirus and tries to delete every .dll and .bat files.  Now if someone clicks on the .bat or .exe file this batch file will run and execute the operations written in the code. So beware of running this batch file or you may need for a restoration on your computer